backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Riad Elimemmedov <riad.elimemmedov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
boto3 = "^1.34.32"
httpx = "^0.26.0"
coverage = "^7.4.1"
django-cleanup = "^8.1.0"
django-jet-reboot = "^1.3.7"
django-money = "^3.4.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
factory-boy = "^3.3.0"
faker = "^22.6.0"
# psycopg2 = "^2.9.9"
pillow = "^10.2.0"
python-decouple = "^3.8"
django-extensions = "^3.2.3"
feedparser = "^6.0.11"
pre-commit = "^3.6.0"
pytest-factoryboy = "^2.6.0"
pytest-django = "^4.8.0"
django-colorfield = "^0.11.0"
django-storages = "^1.14.2"
pytest-mock = "^3.12.0"
django-debug-toolbar = "^4.3.0"
django-lifecycle = "^1.1.2"
django-redis = "^5.4.0"
redis = "^5.0.1"
gunicorn = "^21.2.0"
waitress = "^3.0.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
stripe = "5.4.0"
eth-utils = "^4.0.0"
eth-hash = {extras = ["pycryptodome"], version = "^0.7.0"}
pytest-xdist = "^3.5.0"
django-allauth = "0.52.0"
dj-rest-auth = {version = "4.0.0", extras = ["with-social"]}


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
bandit = "^1.7.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
known_first_party = ["config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
